+++ pkg/apis/kai/v1/config_types.go
@@ -28,6 +28,7 @@ const (
 	ConditionTypeDeployed          ConditionType = "Deployed"
 	ConditionTypeAvailable         ConditionType = "Available"
 	ConditionDependenciesFulfilled ConditionType = "DependenciesFulfilled"
+	ConditionTypeReady             ConditionType = "Ready"
 )
 
 type ConditionReason string
@@ -37,8 +38,10 @@ const (
 	Available                  ConditionReason = "available"
 	Reconciled                 ConditionReason = "reconciled"
 	DependenciesFulfilled      ConditionReason = "dependencies_fulfilled"
+	DependenciesMissing        ConditionReason = "dependencies_missing"
 	PrometheusConnected        ConditionReason = "prometheus_connected"
 	PrometheusConnectionFailed ConditionReason = "prometheus_connection_failed"
+	Ready                      ConditionReason = "ready"
 )
 
 // ConfigSpec defines the desired state of Config