+++ pkg/operator/operands/common/common.go
@@ -31,8 +31,13 @@ var KaiServicesForServiceMonitor = []struct {
 	Name     string
 	Port     string
 	JobLabel string
+
+	LabelSelector map[string]string // optional, if not provided, "app": NAME will be used
+	Namespaces    []string          // optional, if not provided, the service will be monitored in the namespace of the KAI config
 }{
-	{"queuecontroller", "metrics", "queuecontroller"},
+	{"queue-controller", "metrics", "queue-controller", nil, nil},
+	{"kube-state-metrics", "http", "kube-state-metrics",
+		map[string]string{"app.kubernetes.io/name": "kube-state-metrics"}, []string{"monitoring", "default"}},
 }
 
 func AllControllersAvailable(
@@ -209,6 +214,25 @@ func AddK8sClientConfigToArgs(k8sClientConfig *kaiv1common.K8sClientConfig, args
 	}
 }
 
+func CheckCRDsAvailable(ctx context.Context, client client.Reader, crdNames ...string) (bool, error) {
+	for _, name := range crdNames {
+		crd := &metav1.PartialObjectMetadata{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       "CustomResourceDefinition",
+				APIVersion: "apiextensions.k8s.io/v1",
+			},
+		}
+		err := client.Get(ctx, types.NamespacedName{Name: name}, crd)
+		if err != nil {
+			if errors.IsNotFound(err) {
+				return false, nil
+			}
+			return false, fmt.Errorf("failed to check CRD %s: %w", name, err)
+		}
+	}
+	return true, nil
+}
+
 func CheckPrometheusCRDsAvailable(ctx context.Context, client client.Reader, targetCRDs ...string) (bool, error) {
 	var names []string
 	for _, targetCRD := range targetCRDs {
@@ -221,26 +245,7 @@ func CheckPrometheusCRDsAvailable(ctx context.Context, client client.Reader, tar
 			names = append(names, targetCRD)
 		}
 	}
-
-	for _, name := range names {
-		crd := &metav1.PartialObjectMetadata{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "CustomResourceDefinition",
-				APIVersion: "apiextensions.k8s.io/v1",
-			},
-		}
-		err := client.Get(ctx, types.NamespacedName{
-			Name: name,
-		}, crd)
-		if err != nil {
-			if errors.IsNotFound(err) {
-				return false, nil
-			}
-			return false, fmt.Errorf("failed to check for Prometheus CRD: %w", err)
-		}
-	}
-
-	return true, nil
+	return CheckCRDsAvailable(ctx, client, names...)
 }
 
 func MergeAffinities(localAffinity *v1.Affinity,