+++ pkg/operator/operands/prometheus/prometheus_test.go
@@ -59,18 +59,11 @@ var _ = Describe("Prometheus", func() {
 		})
 
 		Context("when Prometheus is disabled", func() {
-			It("should return ServiceAccount only", func(ctx context.Context) {
+			It("should return no objects", func(ctx context.Context) {
 				kaiConfig.Spec.Prometheus.Enabled = ptr.To(false)
 				objects, err := prometheus.DesiredState(ctx, fakeKubeClient, kaiConfig)
 				Expect(err).To(BeNil())
-				Expect(len(objects)).To(Equal(1)) // ServiceAccount only
-			})
-
-			It("should return ServiceAccount only when config is nil", func(ctx context.Context) {
-				kaiConfig.Spec.Prometheus = nil
-				objects, err := prometheus.DesiredState(ctx, fakeKubeClient, kaiConfig)
-				Expect(err).To(BeNil())
-				Expect(len(objects)).To(Equal(1)) // ServiceAccount only
+				Expect(len(objects)).To(Equal(0))
 			})
 		})
 
@@ -104,7 +97,7 @@ var _ = Describe("Prometheus", func() {
 			It("should return Prometheus object when Prometheus Operator is installed", func(ctx context.Context) {
 				objects, err := prometheus.DesiredState(ctx, fakeKubeClient, kaiConfig)
 				Expect(err).To(BeNil())
-				Expect(len(objects)).To(Equal(3)) // ServiceAccount, Prometheus, 1 ServiceMonitor
+				Expect(len(objects)).To(Equal(4)) // ServiceAccount, Prometheus, 2 ServiceMonitors
 
 				prometheusObj := test_utils.FindTypeInObjects[*monitoringv1.Prometheus](objects)
 				Expect(prometheusObj).NotTo(BeNil())
@@ -131,7 +124,7 @@ var _ = Describe("Prometheus", func() {
 
 				objects, err := prometheus.DesiredState(ctx, fakeKubeClient, kaiConfig)
 				Expect(err).To(BeNil())
-				Expect(len(objects)).To(Equal(3)) // ServiceAccount, Prometheus, 1 ServiceMonitor
+				Expect(len(objects)).To(Equal(4)) // ServiceAccount, Prometheus, 2 ServiceMonitor
 
 				prometheusObj := test_utils.FindTypeInObjects[*monitoringv1.Prometheus](objects)
 				Expect(prometheusObj).NotTo(BeNil())
@@ -358,20 +351,6 @@ var _ = Describe("prometheusForKAIConfig", func() {
 			Expect(err).To(BeNil())
 			Expect(len(objects)).To(Equal(0))
 		})
-
-		It("should return empty objects list when config is nil", func(ctx context.Context) {
-			kaiConfig.Spec.Prometheus = nil
-			objects, err := prometheusForKAIConfig(ctx, fakeKubeClient, kaiConfig)
-			Expect(err).To(BeNil())
-			Expect(len(objects)).To(Equal(0))
-		})
-
-		It("should return empty objects list when enabled is nil", func(ctx context.Context) {
-			kaiConfig.Spec.Prometheus.Enabled = nil
-			objects, err := prometheusForKAIConfig(ctx, fakeKubeClient, kaiConfig)
-			Expect(err).To(BeNil())
-			Expect(len(objects)).To(Equal(0))
-		})
 	})
 
 	Context("when Prometheus is enabled", func() {
@@ -474,11 +453,11 @@ var _ = Describe("prometheusForKAIConfig", func() {
 			// The function skips Prometheus CR creation and only creates ServiceMonitors
 			Expect(err).To(BeNil())
 			Expect(objects).NotTo(BeNil())
-			Expect(len(objects)).To(Equal(1)) // 1 ServiceMonitor
+			Expect(len(objects)).To(Equal(2)) // 2 ServiceMonitors
 
 			serviceMonitor := test_utils.FindTypeInObjects[*monitoringv1.ServiceMonitor](objects)
 			Expect(serviceMonitor).NotTo(BeNil())
-			Expect((*serviceMonitor).Name).To(Equal("queuecontroller"))
+			Expect((*serviceMonitor).Name).To(Equal("queue-controller"))
 		})
 
 		It("should return empty objects list when ServiceMonitors are disabled", func(ctx context.Context) {